tests/wpt/web-platform-tests/fetch/metadata/sec-fetch-dest/navigation.https.sub.html

<!DOCTYPE html>
<html>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/fetch/metadata/resources/helper.js></script>
<script>
  test(t => {
    let actual = {
      "dest": "{{headers[sec-fetch-dest]}}"
    };

    assert_header_dest_equals(actual, "document");
  }, "This page's top-level navigation.");
</script>
